Permission checks and minor fixes

+++ src/main/java/six/team/backend/controller/EventController.java
@@ -35,12 +35,19 @@ public class EventController {
         private final static Logger logger = Logger.getLogger(EventController.class);
 
     @RequestMapping(produces = MediaType.APPLICATION_JSON_VALUE,method = RequestMethod.GET)
-    public @ResponseBody ResponseEntity<String> allEvents() {
-        LinkedList<EventStore> events = Event.getAll();
-        JSONObject details = new JSONObject();
-        details.put("events",events);
-        return new ResponseEntity<String>(details.toString(), HttpStatus.OK);
-
+    public @ResponseBody ResponseEntity<String> allEvents(HttpServletRequest request,HttpServletResponse res) {
+        UserDAO UD = new UserDAO();
+        String token = request.getHeader("token");
+        if(UD.getUserGroupPermissions(UD.getUserGroup(token),"eventsview")) {
+            LinkedList<EventStore> events = Event.getAll();
+            JSONObject details = new JSONObject();
+            details.put("events", events);
+            return new ResponseEntity<String>(details.toString(), HttpStatus.OK);
+        }else {
+            JSONObject message = new JSONObject();
+            message.put("events", "You are unauthorized to view this content");
+            return new ResponseEntity<String>(message.toString(), HttpStatus.UNAUTHORIZED);
+        }
     }
     @RequestMapping(produces = MediaType.APPLICATION_JSON_VALUE,value = "/insert",method = RequestMethod.POST)
     public @ResponseBody
@@ -62,43 +69,59 @@ public class EventController {
             System.out.println(e);
         }
         event.setPoints_category(request.getHeader("points_category"));
-        boolean isCreated = Event.createEvent(event);
-        JSONObject createObject = new JSONObject();
-        createObject.put("message", "Event: " + request.getHeader("name") + " was successfully created");
-        if(isCreated) {
+        UserDAO UD = new UserDAO();
+        String token = request.getHeader("token");
+        if(UD.getUserGroupPermissions(UD.getUserGroup(token),"eventsadd")) {
+            boolean isCreated = Event.createEvent(event);
+            JSONObject createObject = new JSONObject();
             createObject.put("message", "Event: " + request.getHeader("name") + " was successfully created");
-            return new ResponseEntity<String>(createObject.toString(), HttpStatus.ACCEPTED);
-        }else{
-            createObject.put("message", "Event: " + request.getHeader("name") + " was not successfully created");
-            return new ResponseEntity<String>(createObject.toString(), HttpStatus.FORBIDDEN);
+            if (isCreated) {
+                createObject.put("message", "Event: " + request.getHeader("name") + " was successfully created");
+                return new ResponseEntity<String>(createObject.toString(), HttpStatus.ACCEPTED);
+            } else {
+                createObject.put("message", "Event: " + request.getHeader("name") + " was not successfully created");
+                return new ResponseEntity<String>(createObject.toString(), HttpStatus.FORBIDDEN);
+            }
+        }else {
+            JSONObject message = new JSONObject();
+            message.put("events", "You are unauthorized to add an event");
+            return new ResponseEntity<String>(message.toString(), HttpStatus.UNAUTHORIZED);
         }
     }
 
     @RequestMapping(produces = MediaType.APPLICATION_JSON_VALUE,value = "/{eventid}", method = RequestMethod.GET)
-    public @ResponseBody ResponseEntity<String> getAttr(@PathVariable(value="eventid") String id ) {
-        EventStore event = Event.getEvent(Integer.parseInt(id));
-        if(event.getName()!=null) {
-            JSONObject details = new JSONObject();
-            details.put("id", id);
-            details.put("title" , event.getName());
-            details.put("location", event.getLocation());
-            details.put("venue", event.getVenue());
-            details.put("points" , event.getPoints());
-            details.put("startdate", event.getStartDate());
-            details.put("enddate", event.getEndDate());
-            details.put("description", event.getDescription());
-            details.put("points_category", event.getPoints_category());
-            JSONArray array = new JSONArray();
-            array.put(details);
-            JSONObject eventInfo = new JSONObject();
-            eventInfo.put("event", array);
-            return new ResponseEntity<String>(eventInfo.toString(), HttpStatus.OK);
-        }else{
-            JSONObject eventInfo = new JSONObject();
-            JSONObject notfound = new JSONObject();
-            notfound.put("Message", "Not Found");
-            eventInfo.put("Event: " , notfound );
-            return new ResponseEntity<String>(eventInfo.toString(), HttpStatus.NOT_FOUND);
+    public @ResponseBody ResponseEntity<String> getAttr(@PathVariable(value="eventid") String id, HttpServletRequest request,HttpServletResponse res) {
+        UserDAO UD = new UserDAO();
+        String token = request.getHeader("token");
+        if(UD.getUserGroupPermissions(UD.getUserGroup(token),"eventsview")) {
+            EventStore event = Event.getEvent(Integer.parseInt(id));
+            if (event.getName() != null) {
+                JSONObject details = new JSONObject();
+                details.put("id", id);
+                details.put("title", event.getName());
+                details.put("location", event.getLocation());
+                details.put("venue", event.getVenue());
+                details.put("points", event.getPoints());
+                details.put("startdate", event.getStartDate());
+                details.put("enddate", event.getEndDate());
+                details.put("description", event.getDescription());
+                details.put("points_category", event.getPoints_category());
+                JSONArray array = new JSONArray();
+                array.put(details);
+                JSONObject eventInfo = new JSONObject();
+                eventInfo.put("event", array);
+                return new ResponseEntity<String>(eventInfo.toString(), HttpStatus.OK);
+            } else {
+                JSONObject eventInfo = new JSONObject();
+                JSONObject notfound = new JSONObject();
+                notfound.put("Message", "Not Found");
+                eventInfo.put("Event: ", notfound);
+                return new ResponseEntity<String>(eventInfo.toString(), HttpStatus.NOT_FOUND);
+            }
+        }else {
+            JSONObject message = new JSONObject();
+            message.put("events", "You are unauthorized to view this content");
+            return new ResponseEntity<String>(message.toString(), HttpStatus.UNAUTHORIZED);
         }
     }
 
@@ -113,34 +136,51 @@ public class EventController {
         event.setDescription(request.getHeader("description"));
         event.setStartDate(request.getHeader("startdate"));
         event.setEndDate(request.getHeader("enddate"));
-        SimpleDateFormat formatter = new SimpleDateFormat("yyyy/MM/dd HH:mm:ss");
+        SimpleDateFormat sdf = new SimpleDateFormat("E MMM dd yyyy HH:mm:ss z",Locale.ENGLISH);
         try{
-            event.setOrderStartDate(formatter.parse(request.getHeader("startdate")));
+            Date date = sdf.parse(request.getHeader("startdate"));
+            event.setOrderStartDate(date);
         }catch(ParseException e){
             System.out.println(e);
         }
         event.setPoints_category(request.getHeader("points_category"));
-        boolean isUpdated = Event.updateEvent(event);
-        JSONObject createObject = new JSONObject();
-        if(isUpdated) {
-            createObject.put("message", "Event ID: " + id + " was successfully updated");
-            return new ResponseEntity<String>(createObject.toString(), HttpStatus.ACCEPTED);
-        }else{
-            createObject.put("message", "Event ID: " + id + " was not successfully updated");
-            return new ResponseEntity<String>(createObject.toString(), HttpStatus.FORBIDDEN);
+        UserDAO UD = new UserDAO();
+        String token = request.getHeader("token");
+        if(UD.getUserGroupPermissions(UD.getUserGroup(token),"eventsedit")) {
+            boolean isUpdated = Event.updateEvent(event);
+            JSONObject createObject = new JSONObject();
+            if (isUpdated) {
+                createObject.put("message", "Event ID: " + id + " was successfully updated");
+                return new ResponseEntity<String>(createObject.toString(), HttpStatus.ACCEPTED);
+            } else {
+                createObject.put("message", "Event ID: " + id + " was not successfully updated");
+                return new ResponseEntity<String>(createObject.toString(), HttpStatus.FORBIDDEN);
+            }
+        }else {
+            JSONObject message = new JSONObject();
+            message.put("events", "You are unauthorized to update this content");
+            return new ResponseEntity<String>(message.toString(), HttpStatus.UNAUTHORIZED);
         }
     }
 
     @RequestMapping(produces = MediaType.APPLICATION_JSON_VALUE,value = "/delete/{eventid}", method = RequestMethod.POST)
-    public @ResponseBody ResponseEntity<String> deleteEvent(@PathVariable(value="eventid") String id){
-        boolean isDeleted = Event.deleteEvent(Integer.parseInt(id));
-        JSONObject createObject = new JSONObject();
-        if(isDeleted) {
-            createObject.put("message", "Event ID: " + id + " was successfully deleted");
-            return new ResponseEntity<String>(createObject.toString(), HttpStatus.ACCEPTED);
-        }else{
-            createObject.put("message", "Event ID: " + id + " was not successfully deleted");
-            return new ResponseEntity<String>(createObject.toString(), HttpStatus.FORBIDDEN);
+    public @ResponseBody ResponseEntity<String> deleteEvent(@PathVariable(value="eventid") String id, HttpServletRequest request,HttpServletResponse re) {
+        UserDAO UD = new UserDAO();
+        String token = request.getHeader("token");
+        if(UD.getUserGroupPermissions(UD.getUserGroup(token),"eventsdelete")) {
+            boolean isDeleted = Event.deleteEvent(Integer.parseInt(id));
+            JSONObject createObject = new JSONObject();
+            if (isDeleted) {
+                createObject.put("message", "Event ID: " + id + " was successfully deleted");
+                return new ResponseEntity<String>(createObject.toString(), HttpStatus.ACCEPTED);
+            } else {
+                createObject.put("message", "Event ID: " + id + " was not successfully deleted");
+                return new ResponseEntity<String>(createObject.toString(), HttpStatus.FORBIDDEN);
+            }
+        }else {
+            JSONObject message = new JSONObject();
+            message.put("events", "You are unauthorized to delete this content");
+            return new ResponseEntity<String>(message.toString(), HttpStatus.UNAUTHORIZED);
         }
     }
